requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
pillow>=8.0.0
numpy>=1.21.0

# Queue service
pika>=1.2.0

# Optional dependencies for enhanced functionality
# opencv-python>=4.5.0  # For OpenCV image processing
# basicsr>=1.4.0        # For Real-ESRGAN support
# gfpgan>=1.3.0         # For face enhancement
